requirements.txt
langchain
python-dotenv
openai